@@ -1,15 +1,12 @@
 /**
- * @license Copyright 2019 The Lighthouse Authors. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ * @license
+ * Copyright 2019 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
  */
-'use strict';
 
-/* eslint-env jest */
-
-const SettingsController = require('../../extension/scripts/settings-controller.js');
-const defaultConfig = require('../../../lighthouse-core/config/default-config.js');
-const i18n = require('../../../lighthouse-core/lib/i18n/i18n.js');
+import * as SettingsController from '../../extension/scripts/settings-controller.js';
+import defaultConfig from '../../../core/config/default-config.js';
+import * as format from '../../../shared/localization/format.js';
 
 describe('Lighthouse chrome extension SettingsController', () => {
   it('default categories should be correct', () => {
@@ -17,7 +14,7 @@ describe('Lighthouse chrome extension SettingsController', () => {
       .map(([id, category]) => {
         return {
           id,
-          title: i18n.getFormatted(category.title, 'en-US'),
+          title: format.getFormatted(category.title, 'en-US'),
         };
       });
     expect(SettingsController.DEFAULT_CATEGORIES).toEqual(categories);
