@@ -1,11 +1,10 @@
 /**
- * @license Copyright 2019 The Lighthouse Authors. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ * @license
+ * Copyright 2019 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
  */
-'use strict';
 
-const Audit = require('lighthouse').Audit;
+import {Audit} from 'lighthouse';
 
 /**
  * @fileoverview A fake additional check of the robots.txt file.
@@ -43,4 +42,4 @@ class PreloadAsAudit extends Audit {
   }
 }
 
-module.exports = PreloadAsAudit;
+export default PreloadAsAudit;
