@@ -1,17 +1,18 @@
 /**
- * @license Copyright 2016 The Lighthouse Authors. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ * @license
+ * Copyright 2016 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
  */
-'use strict';
 
-const debug = require('debug');
-const marky = require('marky');
+import process from 'process';
+import {EventEmitter} from 'events';
 
-const EventEmitter = require('events').EventEmitter;
+import debug from 'debug';
+import * as marky from 'marky';
+
 const isWindows = process.platform === 'win32';
 
-// process.browser is set when browserify'd via the `process` npm module
+// @ts-expect-error: process.browser is set via Rollup.
 const isBrowser = process.browser;
 
 const colors = {
@@ -60,7 +61,11 @@ class Log {
     log(...argsArray);
   }
 
+  /**
+   * @param {string} title
+   */
   static loggerfn(title) {
+    title = `LH:${title}`;
     let log = loggersByTitle[title];
     if (!log) {
       log = debug(title);
@@ -82,16 +87,19 @@ class Log {
     level_ = level;
     switch (level) {
       case 'silent':
-        debug.enable('-*');
+        debug.enable('-LH:*');
         break;
       case 'verbose':
-        debug.enable('*');
+        debug.enable('LH:*');
+        break;
+      case 'warn':
+        debug.enable('-LH:*, LH:*:warn, LH:*:error');
         break;
       case 'error':
-        debug.enable('-*, *:error');
+        debug.enable('-LH:*, LH:*:error');
         break;
       default:
-        debug.enable('*, -*:verbose');
+        debug.enable('LH:*, -LH:*:verbose');
     }
   }
 
@@ -118,30 +126,54 @@ class Log {
     return level_ === 'verbose';
   }
 
+  /**
+   * @param {{msg: string, id: string, args?: any[]}} status
+   * @param {string} level
+   */
   static time({msg, id, args = []}, level = 'log') {
     marky.mark(id);
     Log[level]('status', msg, ...args);
   }
 
+  /**
+   * @param {{msg: string, id: string, args?: any[]}} status
+   * @param {string} level
+   */
   static timeEnd({msg, id, args = []}, level = 'verbose') {
     Log[level]('statusEnd', msg, ...args);
     marky.stop(id);
   }
 
+  /**
+   * @param {string} title
+   * @param {...any} args
+   */
   static log(title, ...args) {
     Log.events.issueStatus(title, args);
     return Log._logToStdErr(title, args);
   }
 
+  /**
+   * @param {string} title
+   * @param {...any} args
+   */
   static warn(title, ...args) {
     Log.events.issueWarning(title, args);
     return Log._logToStdErr(`${title}:warn`, args);
   }
 
+  /**
+   * @param {string} title
+   * @param {...any} args
+   */
   static error(title, ...args) {
     return Log._logToStdErr(`${title}:error`, args);
   }
 
+  /**
+   * @param {string} title
+   * @param {...any} args
+   */
   static verbose(title, ...args) {
     Log.events.issueStatus(title, args);
     return Log._logToStdErr(`${title}:verbose`, args);
@@ -231,11 +263,19 @@ class Log {
 }
 
 Log.events = new Emitter();
+
+/**
+ * @return {PerformanceEntry[]}
+ */
 Log.takeTimeEntries = () => {
   const entries = marky.getEntries();
   marky.clear();
   return entries;
 };
+
+/**
+ * @return {PerformanceEntry[]}
+ */
 Log.getTimeEntries = () => marky.getEntries();
 
-module.exports = Log;
+export default Log;
