@@ -1,158 +1,145 @@
 /**
- * @license Copyright 2018 The Lighthouse Authors. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ * @license
+ * Copyright 2018 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
  */
 
-import _NetworkNode = require('../lighthouse-core/lib/dependency-graph/network-node');
-import _CPUNode = require('../lighthouse-core/lib/dependency-graph/cpu-node');
-import _Simulator = require('../lighthouse-core/lib/dependency-graph/simulator/simulator');
-import Driver = require('../lighthouse-core/gather/driver');
-import ExecutionContext = require('../lighthouse-core/gather/driver/execution-context');
+import {Protocol as Crdp} from 'devtools-protocol/types/protocol.js';
+import {ProtocolMapping as CrdpMappings} from 'devtools-protocol/types/protocol-mapping.js';
 
-declare global {
-  module LH.Gatherer {
-    /** The Lighthouse wrapper around a raw CDP session. */
-    export interface FRProtocolSession {
-      hasNextProtocolTimeout(): boolean;
-      getNextProtocolTimeout(): number;
-      setNextProtocolTimeout(ms: number): void;
-      on<TEvent extends keyof LH.CrdpEvents>(event: TEvent, callback: (...args: LH.CrdpEvents[TEvent]) => void): void;
-      once<TEvent extends keyof LH.CrdpEvents>(event: TEvent, callback: (...args: LH.CrdpEvents[TEvent]) => void): void;
-      addProtocolMessageListener(callback: (payload: LH.Protocol.RawEventMessage) => void): void
-      removeProtocolMessageListener(callback: (payload: LH.Protocol.RawEventMessage) => void): void
-      off<TEvent extends keyof LH.CrdpEvents>(event: TEvent, callback: (...args: LH.CrdpEvents[TEvent]) => void): void;
-      sendCommand<TMethod extends keyof LH.CrdpCommands>(method: TMethod, ...params: LH.CrdpCommands[TMethod]['paramsType']): Promise<LH.CrdpCommands[TMethod]['returnType']>;
+import {ExecutionContext} from '../core/gather/driver/execution-context.js';
+import {NetworkMonitor} from '../core/gather/driver/network-monitor.js';
+import {Fetcher} from '../core/gather/fetcher.js';
+import {ArbitraryEqualityMap} from '../core/lib/arbitrary-equality-map.js';
+
+import {Artifacts, BaseArtifacts, GathererArtifacts, DevtoolsLog, Trace} from './artifacts.js';
+import Config from './config.js';
+import Result from './lhr/lhr.js';
+import Protocol from './protocol.js';
+import Puppeteer from './puppeteer.js';
+import * as Lantern from '../core/lib/lantern/lantern.js';
+
+type CrdpEvents = CrdpMappings.Events;
+type CrdpCommands = CrdpMappings.Commands;
+
+declare module Gatherer {
+  /** The Lighthouse wrapper around a raw CDP session. */
+  interface ProtocolSession {
+    setTargetInfo(targetInfo: Crdp.Target.TargetInfo): void;
+    hasNextProtocolTimeout(): boolean;
+    getNextProtocolTimeout(): number;
+    setNextProtocolTimeout(ms: number): void;
+    on<TEvent extends keyof CrdpEvents>(event: TEvent, callback: (...args: CrdpEvents[TEvent]) => void): void;
+    once<TEvent extends keyof CrdpEvents>(event: TEvent, callback: (...args: CrdpEvents[TEvent]) => void): void;
+    off<TEvent extends keyof CrdpEvents>(event: TEvent, callback: (...args: CrdpEvents[TEvent]) => void): void;
+    sendCommand<TMethod extends keyof CrdpCommands>(method: TMethod, ...params: CrdpCommands[TMethod]['paramsType']): Promise<CrdpCommands[TMethod]['returnType']>;
+    sendCommandAndIgnore<TMethod extends keyof CrdpCommands>(method: TMethod, ...params: CrdpCommands[TMethod]['paramsType']): Promise<void>;
+    dispose(): Promise<void>;
+    onCrashPromise(): Promise<never>;
+  }
+
+  interface Driver {
+    defaultSession: ProtocolSession;
+    executionContext: ExecutionContext;
+    fetcher: Fetcher;
+    url: () => Promise<string>;
+    targetManager: {
+      rootSession(): ProtocolSession;
+      mainFrameExecutionContexts(): Array<Crdp.Runtime.ExecutionContextDescription>;
+      on(event: 'protocolevent', callback: (payload: Protocol.RawEventMessage) => void): void
+      off(event: 'protocolevent', callback: (payload: Protocol.RawEventMessage) => void): void
+    };
+    networkMonitor: NetworkMonitor;
+  }
+
+  interface Context<TDependencies extends DependencyKey = DefaultDependenciesKey> {
+    /** The gather mode Lighthouse is currently in. */
+    gatherMode: GatherMode;
+    /** The connection to the page being analyzed. */
+    driver: Driver;
+    /** The Puppeteer page handle. */
+    page: Puppeteer.Page;
+    /** The set of base artifacts that are always collected. */
+    baseArtifacts: BaseArtifacts;
+    /** The cached results of computed artifacts. */
+    computedCache: Map<string, ArbitraryEqualityMap>;
+    /** The set of available dependencies requested by the current gatherer. */
+    dependencies: Pick<GathererArtifacts, Exclude<TDependencies, DefaultDependenciesKey>>;
+    /** The settings used for gathering. */
+    settings: Config.Settings;
+  }
+
+  interface GatherResult {
+    artifacts: Artifacts;
+    runnerOptions: {
+      resolvedConfig: Config.ResolvedConfig;
+      computedCache: Map<string, ArbitraryEqualityMap>
     }
+  }
 
-    /** The limited driver interface shared between pre and post Fraggle Rock Lighthouse. */
-    export interface FRTransitionalDriver {
-      defaultSession: FRProtocolSession;
-      executionContext: ExecutionContext;
-    }
+  interface LoadData {
+    networkRecords: Array<Artifacts.NetworkRequest>;
+    devtoolsLog: DevtoolsLog;
+    trace?: Trace;
+  }
 
-    /** The limited context interface shared between pre and post Fraggle Rock Lighthouse. */
-    export interface FRTransitionalContext<TDependencies extends DependencyKey = DefaultDependenciesKey> {
-      /** The URL of the page that is currently active. Might be `about:blank` in the before phases */
-      url: string;
-      /** The gather mode Lighthouse is currently in. */
-      gatherMode: GatherMode;
-      /** The connection to the page being analyzed. */
-      driver: FRTransitionalDriver;
-      /** The set of available dependencies requested by the current gatherer. */
-      dependencies: TDependencies extends DefaultDependenciesKey ?
-        {} :
-        Pick<GathererArtifacts, Exclude<TDependencies, DefaultDependenciesKey>>;
-    }
+  type PhaseResultNonPromise = void | GathererArtifacts[keyof GathererArtifacts];
+  type PhaseResult = PhaseResultNonPromise | Promise<PhaseResultNonPromise>
 
-    export interface PassContext {
-      gatherMode: 'navigation';
-      /** The url of the currently loaded page. If the main document redirects, this will be updated to keep track. */
-      url: string;
-      driver: Driver;
-      passConfig: Config.Pass
-      settings: Config.Settings;
-      /** Gatherers can push to this array to add top-level warnings to the LHR. */
-      LighthouseRunWarnings: Array<string | IcuMessage>;
-      baseArtifacts: BaseArtifacts;
-    }
+  type GatherMode = Result.GatherMode;
 
-    export interface LoadData {
-      networkRecords: Array<Artifacts.NetworkRequest>;
-      devtoolsLog: DevtoolsLog;
-      trace?: Trace;
-    }
+  type DefaultDependenciesKey = '__none__'
+  type DependencyKey = keyof GathererArtifacts | DefaultDependenciesKey
 
-    export type PhaseArtifact = |
-        LH.GathererArtifacts[keyof LH.GathererArtifacts] |
-      LH.Artifacts['devtoolsLogs'] |
-      LH.Artifacts['traces'] |
-      LH.Artifacts['WebAppManifest'] |
-      LH.Artifacts['InstallabilityErrors'] |
-      LH.Artifacts['Stacks'];
-    export type PhaseResultNonPromise = void|PhaseArtifact
-    export type PhaseResult = PhaseResultNonPromise | Promise<PhaseResultNonPromise>
+  interface GathererMetaNoDependencies {
+    /**
+     * Used to validate the dependency requirements of gatherers.
+     * If this property is not defined, this gatherer cannot be the dependency of another. */
+    symbol?: Symbol;
+    /** Lists the modes in which this gatherer can run. */
+    supportedModes: Array<GatherMode>;
+  }
 
-    export type GatherMode = 'snapshot'|'timespan'|'navigation';
+  interface GathererMetaWithDependencies<
+    TDependencies extends Exclude<DependencyKey, DefaultDependenciesKey>
+  > extends GathererMetaNoDependencies {
+    /**
+     * The set of required dependencies that this gatherer needs before it can compute its results.
+     */
+    dependencies: Record<TDependencies, Symbol>;
+  }
 
-    export type DefaultDependenciesKey = '__none__'
-    export type DependencyKey = keyof GathererArtifacts | DefaultDependenciesKey
+  type GathererMeta<TDependencies extends DependencyKey = DefaultDependenciesKey> =
+    [TDependencies] extends [DefaultDependenciesKey] ?
+      GathererMetaNoDependencies :
+      GathererMetaWithDependencies<Exclude<TDependencies, DefaultDependenciesKey>>;
 
-    interface GathererMetaNoDependencies {
-      /**
-       * Used to validate the dependency requirements of gatherers.
-       * If this property is not defined, this gatherer cannot be the dependency of another. */
-      symbol?: Symbol;
-      /** Lists the modes in which this gatherer can run. */
-      supportedModes: Array<GatherMode>;
-    }
+  type GatherPhase = keyof Omit<Gatherer.GathererInstance, 'name'|'meta'>
 
-    interface GathererMetaWithDependencies<
-      TDependencies extends DependencyKey = DefaultDependenciesKey
-    > extends GathererMetaNoDependencies {
-      /**
-       * The set of required dependencies that this gatherer needs before it can compute its results.
-       */
-      dependencies: Record<TDependencies, Symbol>;
-    }
+  interface GathererInstance<TDependencies extends DependencyKey = DefaultDependenciesKey> {
+    meta: GathererMeta<TDependencies>;
+    startInstrumentation(context: Context<DefaultDependenciesKey>): Promise<void>|void;
+    startSensitiveInstrumentation(context: Context<DefaultDependenciesKey>): Promise<void>|void;
+    stopSensitiveInstrumentation(context: Context<DefaultDependenciesKey>): Promise<void>|void;
+    stopInstrumentation(context: Context<DefaultDependenciesKey>): Promise<void>|void;
+    getArtifact(context: Context<TDependencies>): PhaseResult;
+  }
 
-    export type GathererMeta<TDependencies extends DependencyKey = DefaultDependenciesKey> =
-      TDependencies extends DefaultDependenciesKey ?
-        GathererMetaNoDependencies :
-        GathererMetaWithDependencies<TDependencies>;
+  type GathererInstanceExpander<TDependencies extends Gatherer.DependencyKey> =
+    // Lack of brackets intentional here to convert to the union of all individual dependencies.
+    TDependencies extends Gatherer.DefaultDependenciesKey ?
+      GathererInstance<Gatherer.DefaultDependenciesKey> :
+      GathererInstance<Exclude<TDependencies, DefaultDependenciesKey>>
+  type AnyGathererInstance = GathererInstanceExpander<Gatherer.DependencyKey>
 
-    export interface GathererInstance {
-      name: keyof LH.GathererArtifacts;
-      beforePass(context: LH.Gatherer.PassContext): PhaseResult;
-      pass(context: LH.Gatherer.PassContext): PhaseResult;
-      afterPass(context: LH.Gatherer.PassContext, loadData: LH.Gatherer.LoadData): PhaseResult;
-    }
-
-    export interface FRGathererInstance<TDependencies extends DependencyKey = DefaultDependenciesKey> {
-      name: keyof LH.GathererArtifacts; // temporary COMPAT measure until artifact config support is available
-      meta: GathererMeta<TDependencies>;
-      beforeNavigation(context: FRTransitionalContext<DefaultDependenciesKey>): Promise<void>|void;
-      beforeTimespan(context: FRTransitionalContext<DefaultDependenciesKey>): Promise<void>|void;
-      afterTimespan(context: FRTransitionalContext<TDependencies>): PhaseResult;
-      afterNavigation(context: FRTransitionalContext<TDependencies>): PhaseResult;
-      snapshot(context: FRTransitionalContext<TDependencies>): PhaseResult;
-    }
-
-    namespace Simulation {
-      export type GraphNode = import('../lighthouse-core/lib/dependency-graph/base-node').Node;
-      export type GraphNetworkNode = _NetworkNode;
-      export type GraphCPUNode = _CPUNode;
-      export type Simulator = _Simulator;
-
-      export interface MetricCoefficients {
-        intercept: number;
-        optimistic: number;
-        pessimistic: number;
-      }
-
-      export interface Options {
-        rtt?: number;
-        throughput?: number;
-        maximumConcurrentRequests?: number;
-        cpuSlowdownMultiplier?: number;
-        layoutTaskMultiplier?: number;
-        additionalRttByOrigin?: Map<string, number>;
-        serverResponseTimeByOrigin?: Map<string, number>;
-      }
-
-      export interface NodeTiming {
-        startTime: number;
-        endTime: number;
-        duration: number;
-      }
-
-      export interface Result {
-        timeInMs: number;
-        nodeTimings: Map<GraphNode, NodeTiming>;
-      }
-    }
+  namespace Simulation {
+    type GraphNode = Lantern.Graph.Node<Artifacts.NetworkRequest>;
+    type GraphNetworkNode = Lantern.Graph.NetworkNode<Artifacts.NetworkRequest>;
+    type GraphCPUNode = Lantern.Graph.CPUNode<Artifacts.NetworkRequest>;
+    type Simulator = Lantern.Simulation.Simulator<Artifacts.NetworkRequest>;
+    type NodeTiming = Lantern.Types.Simulation.NodeTiming;
+    type Result = Lantern.Simulation.Result<Artifacts.NetworkRequest>;
   }
 }
 
-// empty export to keep file a module
-export {};
+export default Gatherer;
