@@ -1,9 +1,16 @@
 /**
- * @license Copyright 2019 The Lighthouse Authors. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ * @license
+ * Copyright 2019 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
  */
-'use strict';
+
+const BACKENDS = [{
+  id: 'psi',
+  title: 'PSI Frontend (pagespeed.web.dev)',
+}, {
+  id: 'viewer',
+  title: 'Lighthouse Viewer (googlechrome.github.io)',
+}];
 
 // Manually define the default categories, instead of bundling a lot of i18n code.
 const DEFAULT_CATEGORIES = [{
@@ -18,12 +25,9 @@ const DEFAULT_CATEGORIES = [{
 }, {
   id: 'seo',
   title: 'SEO',
-}, {
-  id: 'pwa',
-  title: 'Progressive Web App',
 }];
 
-/** @typedef {{selectedCategories: string[], device: string}} Settings */
+/** @typedef {{backend: string, selectedCategories: string[], device: string, locale: string}} Settings */
 
 const STORAGE_KEYS = {
   Categories: 'lighthouse_audits',
@@ -50,6 +54,11 @@ function saveSettings(settings) {
 
   // Stash device setting.
   storage[STORAGE_KEYS.Settings].device = settings.device;
+
+  // Stash backend setting.
+  storage[STORAGE_KEYS.Settings].backend = settings.backend;
+
+  storage[STORAGE_KEYS.Settings].locale = settings.locale;
 
   // Save object to chrome local storage.
   chrome.storage.local.set(storage);
@@ -82,14 +91,17 @@ function loadSettings() {
       const savedSettings = {...defaultSettings, ...result[STORAGE_KEYS.Settings]};
 
       resolve({
+        backend: savedSettings.backend ?? 'psi',
         device: savedSettings.device,
+        locale: savedSettings.locale ?? navigator.language,
         selectedCategories: Object.keys(savedCategories).filter(cat => savedCategories[cat]),
       });
     });
   });
 }
 
-module.exports = {
+export {
+  BACKENDS,
   DEFAULT_CATEGORIES,
   STORAGE_KEYS,
   saveSettings,
